Improve error stack traces.
Rather than using homegrown stack traces allow the use of Devel::StackTrace
or even better Devel::StackTrace::WithLexicals.  Of course allow this to be
controlled via $Search::Elasticsearch::StackTrace for if a trace should be
created $Search::ElasticSearch::StackTracePackage for the package to use
and $Search::ElasticSearch::StackTraceOptions to pass options to the
constructor.