Add support for EC2Search CxnPool
This connection pool module adds the ability for the list of nodes to be
returned from the search of EC2 instances at Amazon EC2.  It makes it so
you can specify a filter (such as instances tagged with type set to
search).  It makes it so you no longer need to have another way to
discover the initial set of nodes.  Of course this module caches the
results for a certain set of time.

The ES server code does this with the AWS plugin to discover other nodes
in the cluster.

There is also the ability to change the returned results, so if you're
running https in front of ES you can change the URI passed for the
nodes.  By default it just returns the instance's hostname and port
9200.

The code supports IAM roles as well so no authentication keys or tokens
need to be supplied if you're using one on the instance.  Yay, security!