Add fork protection to Scroll and Async::Scroll
An existing Scroll instance should not be usable across forks. With this
change, an attempt to do a scroll request in a different process throws
an Illegal exception, and no attempt is made to clear an existing scroll
ID in a forked process.

Thanks to @jeteve for writing most of the code.

Closes #80
Closes #79